test: Adding some more tests
