+++ seagliderOG1/tools.py
@@ -26,6 +26,53 @@ variables_sensors = {
     "PRES_ADCP": "ADVs and turbulence probes",
 }
 
+
+# EFW 2025-01-01: This could be improved
+# Sensor information is contained in the basestation file as a variable (e.g., 'wlbb2f') or
+# in the sg_cal strings, e.g. calibcomm_oxygen or calibcomm_optode contains 'SBE 43F' or 'Optode 4330F'
+# Using this, and the presense of calibration coefficients (e.g., optode_FoilCoefA1), we can make
+# a guess as to the sensor type
+def gather_sensor_info(ds_other, ds_sgcal, firstrun=False):
+    """
+    Gathers sensor information from the provided datasets and organizes it into a new dataset.
+    Parameters:
+    ds_other (xarray.Dataset): The dataset containing sensor data.
+    ds_sgcal (xarray.Dataset): The dataset containing calibration data.
+    firstrun (bool, optional): A flag indicating if this is the first run. Defaults to False.
+    Returns:
+    xarray.Dataset: A dataset containing the gathered sensor information.
+    Notes:
+    - The function looks for specific sensor names in the `ds_other` dataset and adds them to a new dataset `ds_sensor`.
+    - If 'aanderaa4330_instrument_dissolved_oxygen' is present in `ds_other`, it is renamed to 'aa4330'.
+    - If 'Pcor' is present in `ds_sgcal`, an additional sensor 'sbe43' is created based on 'sbe41' with specific attributes.
+    - If 'optode_FoilCoefA1' is present in `ds_sgcal`, an additional sensor 'aa4831' is created based on 'sbe41' with specific attributes.
+    - The function sets appropriate attributes for the sensors 'aa4330', 'aa4831', and 'sbe43' if they are present.
+    """
+
+    # Gather sensors
+    sensor_names = ['wlbb2f', 'sbe41', 'aa4330', 'aa4381', 'aa4330f', 'aa4381f']
+
+    ds_sensor = xr.Dataset()
+    if 'aanderaa4330_instrument_dissolved_oxygen' in ds_other.variables:
+        ds_other['aa4330'] = ds_other['aanderaa4330_instrument_dissolved_oxygen']
+    for sensor in sensor_names:
+        if sensor in ds_other:
+            ds_sensor[sensor] = ds_other[sensor]
+
+    if 'Pcor' in ds_sgcal:
+        ds_sensor['sbe43'] = ds_sensor['sbe41']
+        ds_sensor['sbe43'].attrs['long_name'] = 'Sea-Bird SBE 43F Oxygen Sensor'
+        ds_sensor['sbe43'].attrs['ancillary_variables'] = 'sg_cal_Pcor sg_cal_Foffset sg_cal_A sg_cal_B sg_cal_C sg_cal_E sg_cal_Soc'
+
+    aanderaa_ancillary = 'optode_FoilCoefA1 optode_FoilCoefA2	optode_FoilCoefA3 optode_FoilCoefA4	optode_FoilCoefA5 optode_FoilCoefA6 optode_FoilCoefA7 optode_FoilCoefA8 optode_FoilCoefA9 optode_FoilCoefA10 optode_FoilCoefA11	optode_FoilCoefA12 optode_FoilCoefA13 optode_FoilCoefB1	optode_FoilCoefB2 optode_FoilCoefB3	optode_FoilCoefB4 optode_FoilCoefB5	 optode_FoilCoefB6 optode_PhaseCoef0 optode_PhaseCoef1 optode_PhaseCoef2 optode_PhaseCoef3 optode_ConcCoef0 optode_ConcCoef1 optode_SVU_enabled optode_TempCoef0 optode_TempCoef1 optode_TempCoef2 optode_TempCoef3 optode_TempCoef4 optode_TempCoef5'
+    if 'optode_FoilCoefA1' in sg_cal:
+        ds_sensor['aa4831'] = ds_sensor['sbe41']
+        ds_sensor['aa4831'].attrs['long_name'] = 'Aanderaa 4831F Oxygen Sensor'
+        ds_sensor['aa4831'].attrs['ancillary_variables'] = aanderaa_ancillary
+    if 'aa4330' in ds_sensor.variables:
+        ds_sensor['aa4330'].attrs['long_name'] = 'Aanderaa 4330 Oxygen Sensor'
+        ds_sensor['aa4330'].attrs['ancillary_variables'] = aanderaa_ancillary
+
 def add_sensor_to_dataset(dsa, ds, sg_cal, firstrun=False):
     sensors = list(ds)
     sensor_name_type = {}
@@ -188,17 +235,33 @@ def add_sensor_to_dataset(dsa, ds, sg_cal, firstrun=False):
 
 
 
-def assign_profile_number(ds, divenum_str = 'divenum'):
+def add_dive_number(ds, dive_number=None):
+    """
+    Add dive number as a variable to the dataset. Assumes present in the basestation attributes.
+
+    Parameters:
+    ds (xarray.Dataset): The dataset to which the dive number will be added.
+
+    Returns:
+    xarray.Dataset: The dataset with the dive number added.
+    """
+    if dive_number==None:
+        dive_number = ds.attrs.get('dive_number', np.nan)
+    return ds.assign(divenum=('N_MEASUREMENTS', [dive_number] * ds.dims['N_MEASUREMENTS']))
+
+
+def assign_profile_number(ds, ds1):
     # Remove the variable dive_num_cast if it exists
     if 'dive_num_cast' in ds.variables:
         ds = ds.drop_vars('dive_num_cast')
     # Initialize the new variable with the same dimensions as dive_num
     ds['dive_num_cast'] = (['N_MEASUREMENTS'], np.full(ds.dims['N_MEASUREMENTS'], np.nan))
 
+    ds = add_dive_number(ds, ds1.attrs['dive_number'])
     # Iterate over each unique dive_num
-    for dive in np.unique(ds[divenum_str]):
+    for dive in np.unique(ds['divenum']):
         # Get the indices for the current dive
-        dive_indices = np.where(ds[divenum_str] == dive)[0]
+        dive_indices = np.where(ds['divenum'] == dive)[0]
         # Find the start and end index for the current dive
         start_index = dive_indices[0]
         end_index = dive_indices[-1]
@@ -217,6 +280,7 @@ def assign_profile_number(ds, divenum_str = 'divenum'):
         # Remove the variable PROFILE_NUMBER if it exists
         if 'PROFILE_NUMBER' in ds.variables:
             ds = ds.drop_vars('PROFILE_NUMBER')
+
         # Assign PROFILE_NUMBER as 2 * dive_num_cast - 1
         ds['PROFILE_NUMBER'] = 2 * ds['dive_num_cast'] - 1
     return ds
@@ -362,19 +426,6 @@ def split_by_unique_dims(ds):
     return {dims: dataset for dims, dataset in unique_dims_datasets.items()}
 
 
-def add_dive_number(ds, dive_number):
-    """
-    Add dive number as a variable to the dataset.
-
-    Parameters:
-    ds (xarray.Dataset): The dataset to which the dive number will be added.
-
-    Returns:
-    xarray.Dataset: The dataset with the dive number added.
-    """
-    if dive_number==None:
-        dive_number = ds.attrs.get('dive_number', np.nan)
-    return ds.assign(divenum=('N_MEASUREMENTS', [dive_number] * ds.dims['N_MEASUREMENTS']))
 
 def convert_units(ds):
     """
@@ -522,38 +573,6 @@ def set_best_dtype(ds):
         f"Space saved by dtype downgrade: {int(100 * (bytes_in - bytes_out) / bytes_in)} %",
     )
     return ds
-#===============================================================================
-# Unused functions
-#===============================================================================
-def encode_times(ds):
-    if "units" in ds.time.attrs.keys():
-        ds.time.attrs.pop("units")
-    if "calendar" in ds.time.attrs.keys():
-        ds.time.attrs.pop("calendar")
-    ds["time"].encoding["units"] = "seconds since 1970-01-01T00:00:00Z"
-    for var_name in list(ds):
-        if "time" in var_name.lower() and not var_name == "time":
-            for drop_attr in ["units", "calendar", "dtype"]:
-                if drop_attr in ds[var_name].attrs.keys():
-                    ds[var_name].attrs.pop(drop_attr)
-            ds[var_name].encoding["units"] = "seconds since 1970-01-01T00:00:00Z"
-    return ds
-
-
-def encode_times_og1(ds):
-    for var_name in ds.variables:
-        if "axis" in ds[var_name].attrs.keys():
-            ds[var_name].attrs.pop("axis")
-        if "time" in var_name.lower():
-            for drop_attr in ["units", "calendar", "dtype"]:
-                if drop_attr in ds[var_name].attrs.keys():
-                    ds[var_name].attrs.pop(drop_attr)
-                if drop_attr in ds[var_name].encoding.keys():
-                    ds[var_name].encoding.pop(drop_attr)
-            if var_name.lower() == "time":
-                ds[var_name].attrs["units"] = "seconds since 1970-01-01T00:00:00Z"
-                ds[var_name].attrs["calendar"] = "gregorian"
-    return ds
 
 
 def set_best_dtype_value(value, var_name):
@@ -605,148 +624,37 @@ def find_best_dtype(var_name, da):
 
 
 
-
-
-
-
-
-def add_standard_global_attrs(ds):
-    date_created = datetime.datetime.now().isoformat().split(".")[0]
-    attrs = {
-        "acknowledgement": "This study used data collected and made freely available by Voice of the Ocean Foundation ("
-        "https://voiceoftheocean.org)",
-        "conventions": "CF-1.11",
-        "institution_country": "SWE",
-        "creator_email": "callum.rollo@voiceoftheocean.org",
-        "creator_name": "Callum Rollo",
-        "creator_type": "Person",
-        "creator_url": "https://observations.voiceoftheocean.org",
-        "date_created": date_created,
-        "date_issued": date_created,
-        "geospatial_lat_max": np.nanmax(ds.LATITUDE),
-        "geospatial_lat_min": np.nanmin(ds.LATITUDE),
-        "geospatial_lat_units": "degrees_north",
-        "geospatial_lon_max": np.nanmax(ds.LONGITUDE),
-        "geospatial_lon_min": np.nanmin(ds.LONGITUDE),
-        "geospatial_lon_units": "degrees_east",
-        "contributor_email": "callum.rollo@voiceoftheocean.org, louise.biddle@voiceoftheocean.org, , , , , , ",
-        "contributor_role_vocabulary": "https://vocab.nerc.ac.uk/collection/W08/",
-        "contributor_role": "Data scientist, PI, Operator, Operator, Operator, Operator, Operator, Operator,",
-        "contributing_institutions": "Voice of the Ocean Foundation",
-        "contributing_institutions_role": "Operator",
-        "contributing_institutions_role_vocabulary": "https://vocab.nerc.ac.uk/collection/W08/current/",
-        "agency": "Voice of the Ocean",
-        "agency_role": "contact point",
-        "agency_role_vocabulary": "https://vocab.nerc.ac.uk/collection/C86/current/",
-        "infoUrl": "https://observations.voiceoftheocean.org",
-        "inspire": "ISO 19115",
-        "institution": "Voice of the Ocean Foundation",
-        "institution_edmo_code": "5579",
-        "keywords": "CTD, Oceans, Ocean Pressure, Water Pressure, Ocean Temperature, Water Temperature, Salinity/Density, "
-        "Conductivity, Density, Salinity",
-        "keywords_vocabulary": "GCMD Science Keywords",
-        "licence": "Creative Commons Attribution 4.0 (https://creativecommons.org/licenses/by/4.0/) This study used data collected and made freely available by Voice of the Ocean Foundation (https://voiceoftheocean.org) accessed from https://erddap.observations.voiceoftheocean.org/erddap/index.html",
-        "disclaimer": "Data, products and services from VOTO are provided 'as is' without any warranty as to fitness for "
-        "a particular purpose.",
-        "references": "Voice of the Ocean Foundation",
-        "source": "Voice of the Ocean Foundation",
-        "sourceUrl": "https://observations.voiceoftheocean.org",
-        "standard_name_vocabulary": "CF Standard Name Table v70",
-        "time_coverage_end": str(np.nanmax(ds.time)).split(".")[0],
-        "time_coverage_start": str(np.nanmin(ds.time)).split(".")[0],
-        "variables": list(ds),
-    }
-    for key, val in attrs.items():
-        if key in ds.attrs.keys():
-            continue
-        ds.attrs[key] = val
+def encode_times(ds):
+    if "units" in ds.time.attrs.keys():
+        ds.time.attrs.pop("units")
+    if "calendar" in ds.time.attrs.keys():
+        ds.time.attrs.pop("calendar")
+    ds["time"].encoding["units"] = "seconds since 1970-01-01T00:00:00Z"
+    for var_name in list(ds):
+        if "time" in var_name.lower() and not var_name == "time":
+            for drop_attr in ["units", "calendar", "dtype"]:
+                if drop_attr in ds[var_name].attrs.keys():
+                    ds[var_name].attrs.pop(drop_attr)
+            ds[var_name].encoding["units"] = "seconds since 1970-01-01T00:00:00Z"
     return ds
 
 
-# Deprecated
-def add_sensors_old(ds, dsa):
-    attrs = ds.attrs
-    sensors = []
-    for key, var in attrs.items():
-        if not isinstance(var, str):
-            continue
-        if "{" not in var:
-            continue
-        if isinstance(eval(var), dict):
-            sensors.append(key)
+def encode_times_og1(ds):
+    for var_name in ds.variables:
+        if "axis" in ds[var_name].attrs.keys():
+            ds[var_name].attrs.pop("axis")
+        if "time" in var_name.lower():
+            for drop_attr in ["units", "calendar", "dtype"]:
+                if drop_attr in ds[var_name].attrs.keys():
+                    ds[var_name].attrs.pop(drop_attr)
+                if drop_attr in ds[var_name].encoding.keys():
+                    ds[var_name].encoding.pop(drop_attr)
+            if var_name.lower() == "time":
+                ds[var_name].attrs["units"] = "seconds since 1970-01-01T00:00:00Z"
+                ds[var_name].attrs["calendar"] = "gregorian"
+    return ds
 
-    sensor_name_type = {}
-    for instr in sensors:
-        if instr in ["altimeter"]:
-            continue
-        attr_dict = eval(attrs[instr])
-        if attr_dict["make_model"] not in vocabularies.sensor_vocabs.keys():
-            _log.error(f"sensor {attr_dict['make_model']} not found")
-            continue
-        var_dict = vocabularies.sensor_vocabs[attr_dict["make_model"]]
-        if "serial" in attr_dict.keys():
-            var_dict["serial_number"] = str(attr_dict["serial"])
-            var_dict["long_name"] += f":{str(attr_dict['serial'])}"
-        for var_name in ["calibration_date", "calibration_parameters"]:
-            if var_name in attr_dict.keys():
-                var_dict[var_name] = str(attr_dict[var_name])
-        da = xr.DataArray(attrs=var_dict)
-        sensor_var_name = f"sensor_{var_dict['sensor_type']}_{var_dict['serial_number']}".upper().replace(
-            " ",
-            "_",
-        )
-        dsa[sensor_var_name] = da
-        sensor_name_type[var_dict["sensor_type"]] = sensor_var_name
 
-    for key, var in attrs.copy().items():
-        if not isinstance(var, str):
-            continue
-        if "{" not in var:
-            continue
-        if isinstance(eval(var), dict):
-            attrs.pop(key)
-    ds.attrs = attrs
-
-    for key, sensor_type in variables_sensors.items():
-        if key in dsa.variables:
-            instr_key = sensor_name_type[sensor_type]
-            dsa[key].attrs["sensor"] = instr_key
-
-    return ds, dsa
-
-# Deprecated - is Seaexplorer specific
-def sensor_sampling_period(glider, mission):
-    # Get sampling period of CTD in seconds for a given glider mission
-    fn = f"/data/data_raw/complete_mission/SEA{glider}/M{mission}/sea{str(glider).zfill(3)}.{mission}.pld1.raw.10.gz"
-    df = pd.read_csv(fn, sep=";", dayfirst=True, parse_dates=["PLD_REALTIMECLOCK"])
-    if "LEGATO_TEMPERATURE" in list(df):
-        df_ctd = df.dropna(subset=["LEGATO_TEMPERATURE"])
-    else:
-        df_ctd = df.dropna(subset=["GPCTD_TEMPERATURE"])
-    ctd_seconds = df_ctd["PLD_REALTIMECLOCK"].diff().median().microseconds / 1e6
-
-    if "AROD_FT_DO" in list(df):
-        df_oxy = df.dropna(subset=["AROD_FT_DO"])
-    else:
-        df_oxy = df.dropna(subset=["LEGATO_CODA_DO"])
-    oxy_seconds = df_oxy["PLD_REALTIMECLOCK"].diff().median().microseconds / 1e6
-    sample_dict = {
-        "glider": glider,
-        "mission": mission,
-        "ctd_period": ctd_seconds,
-        "oxy_period": oxy_seconds,
-    }
-    return sample_dict
-
-# Not used
-def natural_sort(unsorted_list):
-    convert = lambda text: int(text) if text.isdigit() else text.lower()  # noqa: E731
-    alphanum_key = lambda key: [convert(c) for c in re.split("([0-9]+)", key)]  # noqa: E731
-    return sorted(unsorted_list, key=alphanum_key)
-
-# Template - doesn't work - requires mailer.sh
-def mailer(subject, message, recipient="callum.rollo@voiceoftheocean.org"):
-    _log.warning(f"email: {subject}, {message}, {recipient}")
-    subject = subject.replace(" ", "-")
-    send_script = sync_script_dir / "mailer.sh"
-    subprocess.check_call(["/usr/bin/bash", send_script, message, subject, recipient])
+#===============================================================================
+# Unused functions
+#===============================================================================
