+++ tests/test_plotters.py
@@ -1,0 +1,7 @@
+import numpy as np
+import pandas as pd
+import xarray as xr
+from pandas import DataFrame
+
+import matplotlib.pyplot as plt
+from seagliderOG1 import plotters
